+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Ferron language server changelog
 
+## `ferron-language-server` UNRELEASED
+
+**Not yet released**
+
+- Fixed on-save hook not working with Zed, due to strict capability handling.
+
 ## `ferron-language-server` 0.1.0
 
 **Released in July 28, 2026**